Move ObservableRangeCollection and ObservableCollectionProvider
