+++ Mapsui.UI.Forms.Tests/LayerThreadingTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Concurrent;
 using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 using Mapsui.Layers;
+using Mapsui.Nts.Providers;
 using Mapsui.UI.Forms;
 using Mapsui.UI.Objects;
 using NUnit.Framework;
