+++ Mapsui/Utilities/ObservableRangeCollection.cs
@@ -1,21 +1,12 @@
-﻿#if __MAUI__
-namespace Mapsui.UI.Maui.Utils;
-#else
-namespace Mapsui.UI.Forms.Utils;
-#endif
-
-using System;
+﻿using System;
 using System.Collections;
-
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Diagnostics;
+
+namespace Mapsui.Utilities;
 
 /// <summary>
 /// Implementation of a dynamic data collection based on generic Collection&lt;T&gt;,
